Added icons for task view.

+++ PrjctMngmt/PrjctMngmt.WebUI/Controllers/DocumentController.cs
@@ -169,7 +169,10 @@ namespace PrjctMngmt.Controllers
 
                 //delete file from the server
                 FileInfo docFile = new FileInfo(basePath + doc.FileName);
-                docFile.Delete();
+                if (docFile.Exists)
+                {
+                    docFile.Delete();
+                }
 
                 return RedirectToAction("Index");
             }
